Default to disabling encrypted backups
Encrypted backups are poorly implemented (both by TWRP for padding a
password with zeros to achieve an AES key and openaes for using
non-standard CBC) and not to be trusted. Do not give users a false
sense of security by offering this option. Disable by default until a
better solution is available.

Developers can re-enable this option for testing by including
TW_EXCLUDE_ENCRYPTED_BACKUPS := false on their device.

Change-Id: I68734adb5e7ec38a0b55c67d01d7cf4b811abecd